adding products to store/shop.html

+++ store/models.py
@@ -1,12 +1,10 @@
 from django.db import models
 
-from django.db import models
-
 
 class Product(models.Model):
     name = models.CharField(max_length=254, default='')
     description = models.TextField()
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.FloatField()
     image = models.ImageField(upload_to='images')
 
     def __str__(self):
